+++ include/params/concrete.hpp
@@ -18,7 +18,7 @@ struct lvl0param {
         ErrorDistribution::ModularGaussian;
     static constexpr inline double α =
         0.000'092'511'997'467'675'6;  // fresh noise, 2^{-13.4}
-    using T = uint32_t;               // Torus representation
+    using T = uint16_t;               // Torus representation
     static constexpr T μ = 1U << (std::numeric_limits<T>::digits - 3);
     static constexpr uint32_t plain_modulus = 8;
     static constexpr double Δ =
